* Show version count in the TableView of the restore explorer (meaning how many versions of a file have been backed up).
* If a restore is attempted on a file that was recently backed up, the restore may fail because the file hasn't shown up in the gmail account
due to mail delays.  Instead of printing a generic error, we say something like "the file was recently backed up and may not be available
for several hours, please try again later."

* During a backup operation, during the initial phase of scanning all the files in the BackupSpecification (this is the estimate operation),
previously we were only updating the progress dialog with the files that were in need of backup.  This means that if you were asking to
backup a huge directory where only a few files were dirty, you wouldn't get much feedback.  Now we update the progress dialog
every time a file is scanned whether it is dirty or not (only for the estimate).  During backup, of course, we only show the files that
are being backed up.

* Use MD5 digests of file contents to :
	a) Generate GUIDs
	b) Detect more accurately whether files need backup.  Before we'd backup a file if its lastMod time was changed.  Now we check the
		MD5 digest to make sure it has changed.
	c) Allow two files with the same content but different names/locations to share the same email backup message.  So if c:/foo.txt is the
		same as c:/biff/bar/baz/other.cfg, only one email will be sent.  The RevisionIdentifiers do ref counting so we know don't delete the
		email until all refs are gone.

* A new serialization scheme that is smokin' fast and very compact relative to our old DataOutputStream scheme, or netscape IFC, or java
serialization.  Compared to our old scheme, its less than half the size, and is twice as fast.  Compared to IFC, its about 2/3rds the size and about
2.5x faster.  Compared to java serialization its about 1/3rd the size, and about 4x as fast.  It is also much more flexible than our original cheazy
DataOutputStream.  It supports shared references, class versioning, and various optimizations like archiving common strings, writing compact
ints/longs, and optimized references based on whether there are multiple holders of the reference or whether the reference is of fixed type or
not.  See com.toubassi.archive.

* If a backup estimate returns that zero files need a backup, we should immediately say no files needed a backup, rather than doing the backup
which just delays the inevitable).