In my last checkin I fixed a bug with the backup explorer not properly updated.  BTW, to repro this bug, do this:
1. In the file explorer, in the left hand side (directory pane), select and expand a non empty directory.  Check it off so it is in the backup list.
2. In a shell, do a 'mkdir foo' within that directory.
3. Do a refresh (f5) in FileBunker.
4. Foo should appear both in the tree and in the table, and be checked off in both.  Before it was not being checked off.

Anyways, when I fixed the Explorer's refresh to update the checkboxes, things slowed down a lot because I was going back to the data model
get the skinny on the checkstate.  Now if the data model is fast, no problem.  But for the backup explorer it requires getting directory listings,
so there was a 1 second pause when you'd go from the restore tab back to the backup tab.

So now I went to the trouble of figuring out how to update checkboxes efficiently.  This turns out to be a pain because certain methods I need
in TreeViewer/CheckboxTreeViewer are protected, so I had to create subclasses.